pyproject.toml
[project]
name = "nintendo-museum"
version = "1.0.0"
description = "Automated monitor for Nintendo Museum ticket availability with Discord notifications"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "curl-cffi>=0.13.0",
    "discord-webhook>=1.4.1",
    "python-dotenv>=1.1.1",
]
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["nintendo", "museum", "ticket", "monitor", "discord", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

[project.scripts]
run-monitor = "main:main"

[dependency-groups]
dev = ["pyinstaller>=6.12.0,<7"]

[tool.hatch.build.targets.sdist]
include = [
    "src",
    "utils",
]

[tool.hatch.build.targets.wheel]
include = [
    "src",
    "utils",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
